Factoring out Request class
